ml/requirements.txt
# Core Dependencies
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pydantic==2.9.2

# ML & AI
sentence-transformers==3.2.1
faiss-cpu==1.12.0
numpy==1.26.4
pandas==2.2.3

# OCR
pytesseract==0.3.13
Pillow==11.0.0

# Utilities
certifi==2024.8.30
python-dotenv==1.0.1

# Database
sqlite3-python==1.0.0
